Resolve suggestions
