+++ WalletWasabi/WabiSabi/Backend/Rounds/CoinJoinStorage/CoinJoinIdStore.cs
@@ -1,12 +1,13 @@
 using NBitcoin;
 using System.IO;
+using System.Linq;
 using WalletWasabi.Logging;
 
 namespace WalletWasabi.WabiSabi.Backend.Rounds.CoinJoinStorage;
 
 public class CoinJoinIdStore : ICoinJoinIdStore
 {
-	public InMemoryCoinJoinIdStore InMemoryCoinJoinIdStore { get; set; }
+	private InMemoryCoinJoinIdStore InMemoryCoinJoinIdStore { get; set; }
 	public string CoinJoinsFilePath { get; set; }
 	private object FileWriteLock { get; set; } = new();
 
@@ -14,10 +15,10 @@ public class CoinJoinIdStore : ICoinJoinIdStore
 	{
 	}
 
-	public CoinJoinIdStore(string ww2CoinjoinsFilePath)
+	public CoinJoinIdStore(string coinJoinIdStoreFilePath)
 	{
-		CoinJoinsFilePath = ww2CoinjoinsFilePath;
-		InMemoryCoinJoinIdStore = InMemoryCoinJoinIdStore.LoadFromFile(ww2CoinjoinsFilePath);
+		CoinJoinsFilePath = coinJoinIdStoreFilePath;
+		InMemoryCoinJoinIdStore = InMemoryCoinJoinIdStore.LoadFromFile(coinJoinIdStoreFilePath);
 	}
 
 	public void Append(uint256 id)
@@ -39,5 +40,24 @@ public class CoinJoinIdStore : ICoinJoinIdStore
 	public bool Contains(uint256 id)
 	{
 		return InMemoryCoinJoinIdStore.Contains(id);
+	}
+
+	public static void ImportWW1CoinJoinsToWW2(string ww1CoinJoinsFilePath, string coinJoinIdStoreFilePath)
+	{
+		try
+		{
+			var oldCoinjoins = File.ReadAllLines(ww1CoinJoinsFilePath);
+
+			var newCoinjoins = File.ReadAllLines(coinJoinIdStoreFilePath);
+			var missingOldCoinjoins = oldCoinjoins.Except(newCoinjoins);
+			if (missingOldCoinjoins.Any())
+			{
+				File.AppendAllLines(coinJoinIdStoreFilePath, missingOldCoinjoins);
+			}
+		}
+		catch (Exception exc)
+		{
+			Logger.LogError("Failed to import old coinjoins. Reason:", exc);
+		}
 	}
 }
