+++ WalletWasabi/WabiSabi/Backend/Rounds/CoinJoinStorage/InMemoryCoinJoinIdStore.cs
@@ -41,27 +41,4 @@ public class InMemoryCoinJoinIdStore
 		var store = new InMemoryCoinJoinIdStore(lines);
 		return store;
 	}
-
-	public void ImportWW1CoinJoinsToWW2(string ww1CoinJoinsFilePath, string coinJoinIdStoreFilePath)
-	{
-		try
-		{
-			var oldCoinjoins = File.ReadAllLines(ww1CoinJoinsFilePath);
-
-			var newCoinjoins = File.ReadAllLines(coinJoinIdStoreFilePath);
-			var missingOldCoinjoins = oldCoinjoins.Except(newCoinjoins);
-			if (missingOldCoinjoins.Any())
-			{
-				foreach (var coinjoinId in missingOldCoinjoins)
-				{
-					Add(uint256.Parse(coinjoinId));
-				}
-				File.AppendAllLines(coinJoinIdStoreFilePath, missingOldCoinjoins);
-			}
-		}
-		catch (Exception exc)
-		{
-			Logger.LogError("Failed to import old coinjoins. Reason:", exc);
-		}
-	}
 }
